Added LineID test utility: line-id-check. Removed LineID route validation test.
