+++ tests/tfl.rs
@@ -66,32 +66,6 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn confirm_all_line_ids() {
-        let client = get_client();
-
-        let mut lids = HashMap::new();
-        for l in linemodels::LineID::iter() {
-            lids.insert(l.line().to_lowercase(), l);
-        }
-
-        let routes = client.routes().fetch().await.expect("Cannot fetch routes");
-
-        for l in routes {
-            // println!("Line: {} -> {}", l.name, l.id);
-            match lids.remove(&l.id.to_lowercase()) {
-                Some(r) => {
-                    assert_eq!(l.name, r.to_string());
-                }
-                // None => panic!("Missing line: {}", l.id),
-                None => println!("Missing line: {}", l.id),
-            }
-        }
-
-        // assert!(lids.is_empty(), "Extra lines defined: {:?}", lids);
-        println!("Extra lines: {:?}", lids);
-    }
-
-    #[tokio::test]
     async fn it_fetches_disruptions_by_mode() {
         let client = get_client();
